chore(ui): delete orphaned legacy deviceConfig Pinia store
`stores/deviceConfig.ts` wired nine device-config API functions but was
only consumed by `DeviceConfigHistoryPage` (two actions: `saveStoredConfig`,
`fetchHistory`). The main `DeviceConfigPage` moved to the template/override
paradigm via `stores/deviceConfigNew.ts` some time ago, leaving the legacy
store as dead weight with a single lingering consumer.

Refactor `DeviceConfigHistoryPage.vue` to call `getConfigHistory` and
`updateDeviceConfig` directly from `@/api/deviceConfig`, managing local
loading/error/history state in the setup block — matching the patterns
used elsewhere where a lightweight page doesn't warrant a Pinia store.

Then delete `ui/src/stores/deviceConfig.ts` entirely.

The API module `ui/src/api/deviceConfig.ts` is unchanged — it remains the
source of truth for device-config API calls.

The two functions named in #130 but not in the codebase (`bulkApplyConfig`,
`getConfigDiff`) are out of scope — neither exists on backend nor frontend;
drop from acceptance criteria.

Closes #130.
